Move houses to a new package
